feat: add Spectrum2DGenerator
