+++ src/types/data2D.d.ts
@@ -1,0 +1,6 @@
+import type { Data } from './data';
+
+interface Data2D extends Data {
+  [index: string]: Float64Array | number[] | Float64Array[] | number[][];
+  z: Float64Array[] | number[][];
+}
