+++ src/types/xyNumber.d.ts
@@ -1,0 +1,5 @@
+export interface xyNumber {
+  [index: string]: number;
+  x: number;
+  y: number;
+}
